Fix up processing tests
