+++ tests/ImageSharp.Web.Tests/TestUtilities/TestServerFixture.cs
@@ -4,11 +4,16 @@
 using System;
 using System.Collections.Generic;
 using System.Net.Http;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using SixLabors.ImageSharp.Web.DependencyInjection;
+using SixLabors.ImageSharp.Web.Middleware;
+using Xunit;
 
 namespace SixLabors.ImageSharp.Web.Tests.TestUtilities
 {
@@ -41,9 +46,67 @@ namespace SixLabors.ImageSharp.Web.Tests.TestUtilities
 
         public IServiceProvider Services { get; private set; }
 
-        protected abstract void ConfigureServices(IServiceCollection services);
+        protected void ConfigureServices(IServiceCollection services)
+        {
+            IImageSharpBuilder builder = services.AddImageSharp(options =>
+            {
+                Func<ImageCommandContext, Task> onParseCommandsAsync = options.OnParseCommandsAsync;
 
-        protected abstract void Configure(IApplicationBuilder app);
+                options.OnParseCommandsAsync = context =>
+                {
+                    Assert.NotNull(context);
+                    Assert.NotNull(context.Context);
+                    Assert.NotNull(context.Commands);
+                    Assert.NotNull(context.Parser);
+
+                    return onParseCommandsAsync.Invoke(context);
+                };
+
+                Func<ImageProcessingContext, Task> onProcessedAsync = options.OnProcessedAsync;
+
+                options.OnProcessedAsync = context =>
+                {
+                    Assert.NotNull(context);
+                    Assert.NotNull(context.Commands);
+                    Assert.NotNull(context.ContentType);
+                    Assert.NotNull(context.Context);
+                    Assert.NotNull(context.Extension);
+                    Assert.NotNull(context.Stream);
+
+                    return onProcessedAsync.Invoke(context);
+                };
+
+                Func<FormattedImage, Task> onBeforeSaveAsync = options.OnBeforeSaveAsync;
+
+                options.OnBeforeSaveAsync = context =>
+                {
+                    Assert.NotNull(context);
+                    Assert.NotNull(context.Format);
+                    Assert.NotNull(context.Encoder);
+                    Assert.NotNull(context.Image);
+
+                    return onBeforeSaveAsync.Invoke(context);
+                };
+
+                Func<HttpContext, Task> onPrepareResponseAsync = options.OnPrepareResponseAsync;
+
+                options.OnPrepareResponseAsync = context =>
+                {
+                    Assert.NotNull(context);
+                    Assert.NotNull(context.Response);
+
+                    return onPrepareResponseAsync.Invoke(context);
+                };
+            })
+            .ClearProviders()
+            .AddProcessor<CacheBusterWebProcessor>();
+
+            this.ConfigureCustomServices(services, builder);
+        }
+
+        protected virtual void Configure(IApplicationBuilder app) => app.UseImageSharp();
+
+        protected abstract void ConfigureCustomServices(IServiceCollection services, IImageSharpBuilder builder);
 
         protected virtual void Dispose(bool disposing)
         {
